refactor: move schema e model para shared/diabetes
