+++ src/app/api/lista-nominal/diabetes/pages/[page]/route.ts
@@ -1,5 +1,5 @@
-import type { CoeqPageRequestBody } from "@/features/acf/diabetes/common/schema";
-import { coeqPageRequestBody as queryParamsSchema } from "@/features/acf/diabetes/common/schema";
+import type { CoeqPageRequestBody } from "@/features/acf/shared/diabetes/schema";
+import { coeqPageRequestBody as queryParamsSchema } from "@/features/acf/shared/diabetes/schema";
 import {
     AuthenticationError,
     decodeToken,
