+++ src/features/acf/backend/diabetes/modules/externalCards/controller.ts
@@ -1,4 +1,4 @@
-import type { AcfDashboardType } from "../../../../diabetes/common/model";
+import type { AcfDashboardType } from "../../../../shared/diabetes/model";
 import type { ExternalCardDataItem } from "../../model";
 import { externalCardsDbToModel } from "./adapter";
 import { externalCardsDataForTeam } from "./repository";
