+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/container.tsx
@@ -1,6 +1,6 @@
 import { MUNICIPIOS } from "@/constants/municipios";
 import { externalCardsAcfDashboardDataControllerForTeam } from "@/features/acf/backend/diabetes/modules/externalCards/controller";
-import type { AcfDashboardType } from "@/features/acf/diabetes/common/model";
+import type { AcfDashboardType } from "@/features/acf/shared/diabetes/model";
 import { getExternalCardsProps } from "@/helpers/cardsList";
 import type { ProfileIdValue } from "@/types/profile";
 import type { CardProps } from "@impulsogov/design-system/dist/molecules/Card/Card";
