+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/logic.ts
@@ -1,11 +1,11 @@
+import type { PanelSelectorWithCardsProps } from "@impulsogov/design-system/dist/organisms/PanelSelectorWithCards/PanelSelectorWithCards";
 import React from "react";
-import type { AcfDashboardType } from "../../../../common/model";
+import type { AcfDashboardType } from "../../../../../shared/diabetes/model";
 import type {
     ExtendedPanelSelectorWithCardsProps,
     ExtendedsubTabsWithChildrenAndChildrenDataProps,
     Tabs,
 } from "./presentation";
-import type { PanelSelectorWithCardsProps } from "@impulsogov/design-system/dist/organisms/PanelSelectorWithCards/PanelSelectorWithCards";
 
 export const subTabChildrenSelector = (
     tabs: Tabs,
