+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/List/container.tsx
@@ -4,7 +4,7 @@ import { SearchToolBar } from "@features/acf/common/frontend/SearchToolBar";
 import { WithFilters } from "@features/acf/common/frontend/WithFilters";
 import { WithSorting } from "@features/acf/common/frontend/WithSorting";
 import React from "react";
-import type { AcfDashboardType } from "../../../../../../common/model";
+import type { AcfDashboardType } from "../../../../../../../shared/diabetes/model";
 import type { AppliedFiltersCoeq } from "./modules/CoeqDataTable";
 import { CoeqDataTable } from "./modules/CoeqDataTable";
 import { CoeqFiltersBar } from "./modules/CoeqFiltersBar";
