+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/List/modules/CoeqFiltersBar/logic.ts
@@ -3,8 +3,8 @@ import type {
     PatientAgeRange,
     PatientStatus,
     VisitantCommunityHealthWorker,
-} from "@/features/acf/diabetes/common/model";
-import type * as schema from "@/features/acf/diabetes/common/schema";
+} from "@/features/acf/shared/diabetes/model";
+import type * as schema from "@/features/acf/shared/diabetes/schema";
 import type * as z from "zod/v4";
 import { nameFormatter } from "../../../../logic";
 import { referenceOrder } from "./consts";
