+++ src/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/modules/List/modules/CoeqFiltersBar/service.ts
@@ -1,5 +1,5 @@
 "use client";
-import type * as schema from "@features/acf/diabetes/common/schema";
+import type * as schema from "@/features/acf/shared/diabetes/schema";
 import type { AxiosResponse } from "axios";
 import axios from "axios";
