+++ src/services/lista-nominal/ListaNominal.ts
@@ -1,8 +1,8 @@
 // import type { DiabetesAcfItem } from "@/features/acf/modules/AcfDashboardPage/modules/PanelSelector/modules/dashboards/modules/TableWithFilters/modules/PaginatedTable/modules/DataTable/modules/diabetes/diabetes.model";
-import type { DiabetesAcfItem } from "@/features/acf/diabetes/common/model";
+import type { DiabetesAcfItem } from "@/features/acf/shared/diabetes/model";
 import type { GridSortDirection } from "@mui/x-data-grid";
-import axios from "axios";
 import type { AxiosResponse } from "axios";
+import axios from "axios";
 
 //TODO: Remover este arquivo
 export interface SortingItem {
