+++ src/utils/subTabChildrenSelector.ts
@@ -1,10 +1,10 @@
 import React from "react";
 
-import type { AcfDashboardType } from "@/features/acf/diabetes/common/model";
 import type {
     ExtendedPanelSelectorWithCardsProps,
     ExtendedsubTabsWithChildrenAndChildrenDataProps,
 } from "@/features/acf/diabetes/frontend/AcfPage/modules/PanelSelector/presentation";
+import type { AcfDashboardType } from "@/features/acf/shared/diabetes/model";
 
 //TODO: Remover daqui
 /**
